requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain-text-splitters
faiss-cpu
groq
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
numpy
requests